Moved xxxUtil files from Slice to respective slice2xxx folders.
