+++ cpp/src/slice2cpp/CPlusPlusUtil.h
