+++ cpp/src/slice2java/JavaUtil.h
