+++ cpp/src/slice2py/msbuild/slice2py.vcxproj
@@ -103,9 +103,14 @@
   </ItemDefinitionGroup>
   <ItemGroup>
     <ClCompile Include="..\Main.cpp" />
+    <ClCompile Include="..\Python.cpp" />
+    <ClCompile Include="..\PythonUtil.cpp" />
   </ItemGroup>
   <ItemGroup>
     <ResourceCompile Include="..\Slice2Py.rc" />
+  </ItemGroup>
+  <ItemGroup>
+    <ClInclude Include="..\PythonUtil.h" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\..\IceUtil\msbuild\iceutil\iceutil.vcxproj">
